test(sub-select): update test case
